adding all the code
