+++ from_scratch_unet/utils.py
@@ -62,6 +62,7 @@ def check_accuracy(loader, model, device="cuda"):
     with torch.no_grad():
         for x, y in loader:
             x = x.to(device)
+            print(f"Feature batch shape: {x.size()}")
             y = y.to(device).unsqueeze(1)
             preds = torch.sigmoid(model(x))
             preds = (preds > 0.5).float()
