--- pretrained_unet/model.py
@@ -1,10 +1,0 @@
-import segmentation_models_pytorch as smp
-
-model = smp.Unet(
-    encoder_name = "resnet34",
-    encoder_weights= "imagenet",
-    in_channels= 3,
-    classes = 1
-)
-print(model)
-
